feat: create DTO layer for User entity

+++ backend/BookerApp/Entities/Book.cs
@@ -2,7 +2,7 @@ namespace BookerApp.Models
 {
     public class Book
     {
-        public int Id {get; set;}
+        public Long Id {get; set;}
         public String Title { get; set;}
         public String Author { get; set;}
         public String Publisher { get; set;}
